fixes inability to access pages >= 256 on the SPI flash on the P1.

+++ platform/MCU/STM32F2xx/SPARK_Firmware_Driver/src/spi_flash.c
@@ -188,7 +188,7 @@ static void sFLASH_WritePage(const uint8_t* pBuffer, uint32_t WriteAddr, uint32_
  */
 void sFLASH_WriteBuffer(const uint8_t* pBuffer, uint32_t WriteAddr, uint32_t NumByteToWrite)
 {
-    uint8_t NumOfPage = 0, NumOfSingle = 0, Addr = 0, count = 0, temp = 0;
+    uint16_t NumOfPage = 0, NumOfSingle = 0, Addr = 0, count = 0, temp = 0;
 
     Addr = WriteAddr % sFLASH_PROGRAM_PAGESIZE;
     count = sFLASH_PROGRAM_PAGESIZE - Addr;
